requirements.txt
tensorflow
numpy
matplotlib
pandas
Pillow
notebook
pyyaml